# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: dLNk Attack Platform API
  description: |
    Comprehensive API documentation for dLNk Attack Platform.
    
    ## Features
    - Authentication & Authorization
    - Target Management
    - Attack Orchestration
    - Agent Control
    - Zero-Day Hunter
    - Real-time Updates
    
    ## Authentication
    All endpoints require JWT token authentication except `/auth/login` and `/auth/register`.
    Include the token in the `Authorization` header: `Bearer <token>`
    
  version: 1.0.0
  contact:
    name: dLNk Security Team
    email: security@dlnk.io
  license:
    name: Proprietary
    url: https://dlnk.io/license
servers:
  - url: https://api.dlnk.io/v1
    description: Production server
  - url: https://staging-api.dlnk.io/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /targets:
    get:
      tags:
        - Targets
      summary: List all targets
      description: Retrieve list of all registered targets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, scanning]
      responses:
        '200':
          description: List of targets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Target'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Targets
      summary: Add new target
      description: Register a new target for scanning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - host
              properties:
                name:
                  type: string
                  example: Production Server
                host:
                  type: string
                  example: 192.168.1.100
                port:
                  type: integer
                  example: 80
                description:
                  type: string
                  example: Main production web server
      responses:
        '201':
          description: Target created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Target'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Target:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        status:
          type: string
          enum: [active, inactive, scanning]
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer